refactor: removed uncessary `#include`s in `sockets-polling.h` refactor: moved function definitions & `#include` that are only used once to their individual files
