+++ include/sockets-polling.h
@@ -1,32 +1,31 @@
 #ifndef SOCKETS_POLLING_h
 #define SOCKETS_POLLING_h
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-#include <netdb.h>
-#include <poll.h>
-#include <vector>
+
+#include <netdb.h> // struct addrinfo, freeaddrinfo (in setupListeningSocket only)
+#include <poll.h> // poll(), struct pollfd
 #include <iostream>
 #include <sstream>
-#include <fcntl.h>
+#include <fcntl.h> // setNonBlocking()
 #include <vector>
-#include <dirent.h> //for opendir, readdir, closedir
-#include <sys/stat.h> // for stat
-#include <algorithm> // for find
+#include <algorithm> // for isListener
+
 #include "Configuration.hpp"
-#include "http-request.h"
-#include "./Buffer.h"
 #include "Client.h"
-#include "http-response.h"
-#include <algorithm>
-#include <cctype>
 #include "Cgi.hpp"
+
+// #include <stdio.h>
+// #include <stdlib.h>
+// #include <string.h> // memset
+// #include <unistd.h>
+// #include <sys/types.h>
+// #include <sys/socket.h>
+// #include <arpa/inet.h> // inet_top
+// #include <dirent.h> //for opendir, readdir, closedir
+// #include <sys/stat.h> // for stat
+// #include "http-request.h"
+// #include "http-response.h"
+// #include <cctype>
 
 #define RESET "\033[0m"
 #define BOLD "\033[1m"
@@ -36,15 +35,18 @@
 #define GREY "\033[90m"
 #define BLUE "\033[0;34m"
 
-
-
 // Setup Listening Socket
 void setupListeningSocket(std::vector<struct pollfd>& pfds, std::vector<int>& listeners, Server& server);
 
-bool isListener(std::vector<int>& listeners, int fd);
-Client* findClientByFd(std::vector<Client>& clients, int fd);
-
+// client
 void acceptClient(std::vector<struct pollfd>& pfds, std::vector<Client>& clients, int listener);
+
+// Utils
+bool isListener(std::vector<int>& listeners, int fd);
+int  setNonBlocking(int fd);
+Client* findClientByFd(std::vector<Client>& clients, int fd);
+std::vector<Client>::iterator disconnectClient(std::vector<Client>& clients, std::vector<Client>::iterator &clientIt, std::vector<struct pollfd>& pfds);
+void clearDisconnectedClients(std::vector<Client>& clients, std::vector<struct pollfd>& pfds);
 
 // handle poll events
 void handlePollIn(std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers,
@@ -54,22 +56,13 @@ void handlePollHup(Client& client);
 void handlePollErr(Client& client);
 
 
-// // Read Request Utils
-// int readRequestHeader(Client& client, std::string& headerStr, const size_t bufferSize);
-// int readRequestBody(Client& client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
-// int receiveClientRequest(Client& client, std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers);
-// int readByChunkedEncoding(Client& client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
+// ! Moved functions that are only used once to their individual files
+// std::string readDirectorytoString(const std::string &directoryPath, const std::string& uri); // used in handleGetRequest only
+// void resolveLocationPath(const std::string& uri, Client& client); // used in handlePollIn() only
 
-// Utils
-int  setNonBlocking(int fd);
-std::vector<Client>::iterator disconnectClient(std::vector<Client>& clients, std::vector<Client>::iterator &clientIt, std::vector<struct pollfd>& pfds);
-void clearDisconnectedClients(std::vector<Client>& clients, std::vector<struct pollfd>& pfds);
 
-// utils2
-void resolveLocationPath(const std::string& uri, Client& client);
-//std::string resolveHttpPath(const std::string& uri, Server &server);
-std::string readDirectorytoString(const std::string &directoryPath, const std::string& uri);
-std::pair<std::string, std::string> getIpAndPortFromSocketFd(int fd);
+// used in recvClientRequest() and timeout functions
+std::pair<std::string, std::string> getIpAndPortFromSocketFd(int fd); 
 Server& getDefaultServerBlockByIpPort(std::pair<std::string, std::string> ipPort, std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers);
 
 // PRINT
