+++ src/read-request/chunked-encoding.cpp
@@ -1,5 +1,6 @@
 #include "read-request.h"
 #include "timeout.h"
+#include <cctype>
 
 static bool doesLineHaveCRLF(const std::string& buffer)
 {
