+++ src/sockets-polling/poll.cpp
@@ -1,15 +1,10 @@
 #include "../include/sockets-polling.h"
 #include "../include/read-request.h"
 
-static void setPfdTrackPollOutOnly(struct pollfd& pfd) {
-    pfd.events &= ~POLLIN;
-    pfd.events |= POLLOUT;
-}
-
-static void setPfdTrackPollInOnly(struct pollfd& pfd) {
-    pfd.events &= ~POLLOUT;
-    pfd.events |= POLLIN;
-}
+static void setPfdTrackPollOutOnly(struct pollfd& pfd);
+static void setPfdTrackPollInOnly(struct pollfd& pfd);
+static void resolveLocationPath(const std::string& uri, Client& client);
+static void sendResponseToClient(int fd, HttpResponse& response);
 
 void handlePollIn(std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers,
                     struct pollfd& pfd, Client& client)
@@ -39,12 +34,6 @@ void handlePollIn(std::map< std::pair<std::string, std::string> , std::vector<Se
     setPfdTrackPollOutOnly(pfd);
 }
 
-static void sendResponseToClient(int fd, HttpResponse& response)
-{
-    send(fd, response.toString().c_str(), response.toString().size(), 0);
-    std::cout << "server: Response sent to client.\n";
-}
-
 void handlePollOut(struct pollfd& pfd, Client& client)
 {
     std::cout << "POLLOUT: socket " << client.fd << '\n';
@@ -71,4 +60,97 @@ void handlePollErr(Client& client)
 {
     std::cout << "POLLERR: socket " << client.fd << '\n';
     client.connState = DISCONNECTED;
+}
+
+static void setPfdTrackPollOutOnly(struct pollfd& pfd) 
+{
+    pfd.events &= ~POLLIN;
+    pfd.events |= POLLOUT;
+}
+
+static void setPfdTrackPollInOnly(struct pollfd& pfd) 
+{
+    pfd.events &= ~POLLOUT;
+    pfd.events |= POLLIN;
+}
+
+/**
+ * @brief normalize the multiple "/" in the relative uri to one "/"
+*/
+static std::string trimMultipleSlash(const std::string &relativeUri)
+{
+    std::string result;
+    bool lastSlash = false;
+    for (size_t i = 0; i < relativeUri.size(); ++i)
+    {
+        if (relativeUri[i] == '/') {
+            if (lastSlash == false)
+            {
+                result += '/';
+                lastSlash = true;
+            }
+        }
+        else
+        {
+            result += relativeUri[i];
+            lastSlash = false;
+        }
+    }
+    return result;
+}
+
+/**
+ * @brief Validate the relative URI and set the location path
+ * @note 1. relativeUri is to identify is there extra value after trim the location path
+ * @note 2. check for "/" for prevent multiple slashes in the path to identify as a directory
+*/
+static void validateRelativeUri(const std::string &relativeUri, Client& client, const std::string &locationType)
+{
+    std::string getRelativeUri = relativeUri.substr(client.location.getLocaPath().length());
+    std::string result = trimMultipleSlash(getRelativeUri);
+    std::cout << "Relative path: " << result << "\n"; ////debug
+    if (result.empty() || result == "/")
+    {
+        std::cout << GREEN "Alias/Root path: " << getFullPath(locationType + result) << "\n" RESET; ////debug
+        if (client.server.getRoot().empty())
+            client.locationPath = getFullPath(locationType + result);
+        else
+            client.locationPath = client.server.getRoot() + locationType + result;
+    }
+    else
+    {
+        std::cout << GREEN "Alias/Root path extra value: " << getFullPath(locationType + result) << "\n" RESET; ////debug
+        if (client.server.getRoot().empty())
+            client.locationPath = getFullPath(locationType + result);
+        else
+            client.locationPath = client.server.getRoot() + locationType + result;
+    }
+}
+
+/**
+ * @brief Get the full path from the alias or root
+*/
+static void resolveLocationPath(const std::string& uri, Client& client)
+{
+    const Location location = client.server.getLocationPath(uri);
+    if (!location.getAlias().empty())
+    {
+        std::cout << "Alias found\n"; ////debug"
+        validateRelativeUri(uri, client, location.getAlias());
+    }
+    else if (!location.getRoot().empty())
+    {
+        std::cout << "Root found\n"; ////debug
+        validateRelativeUri(uri, client, location.getRoot());
+    }
+    else {
+        std::cout << "No root or alias\n";
+        client.locationPath = client.server.getRoot() + uri + "/";
+    }
+}
+
+static void sendResponseToClient(int fd, HttpResponse& response)
+{
+    send(fd, response.toString().c_str(), response.toString().size(), 0);
+    std::cout << "server: Response sent to client.\n";
 }
